+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added option for embolden negative words during simplification via SUMM AI
+
 ### Optimized
 
 - Check for completed setup before initialising any admin options for this plugin